update STAPI parameter to DevDrive (#206)
This PR updates the parameter we give to the WMI Storage Format API from DeveloperVolume to DevDrive. This was a recent change from the storage team and has just hit our dev branches and Windows insider branches. So, I'm updating it here too as without this change the WMI call will fail with "Parameter not found".