Experimental Feature: Quiet Background Processes UI and WinRT server
Adding the "Enable 'Quiet Background Processes'" experimental feature to Dev Home.
It's a toggle that is enabled similarly to Focus Assist mode, that expires after a period of time.
When you enable it/toggle it on via the UI, Dev Home will call into the OS to enable quieting down background services (e.g. Search Indexer).

* Requires 'Admin'
* Requires OS build 26024 or higher